refactor: moves header files for platform independent examples into an include directory
