+++ examples/implementation/platform_independent/SDF_example_008/main_platform_independent.c
@@ -1,8 +1,8 @@
 typedef int token;
 
+#include "include/common.h"
 #include <stdio.h>
 
-#include "common.h"
 /* Netlist */
 
 /*
