+++ examples/implementation/platform_independent/include/actor_templates.h
