+++ examples/implementation/platform_independent/include/buffer_blocking_pc.h
