+++ examples/implementation/platform_independent/include/buffer_nonblocking.h
