+++ examples/implementation/platform_independent/include/common.h
